Move exec plugin to its own file
