+++ exec_plugin.go
@@ -1,0 +1,71 @@
+package main
+
+import (
+	"bytes"
+	"fmt"
+	"io/ioutil"
+	"os"
+
+	"go.mozilla.org/sops/v3/cmd/sops/formats"
+	"go.mozilla.org/sops/v3/decrypt"
+	"sigs.k8s.io/yaml"
+)
+
+type ksops struct {
+	Files []string `json:"files,omitempty" yaml:"files,omitempty"`
+}
+
+// main executes KOSPS as an exec plugin
+func main() {
+	if len(os.Args) != 2 {
+		fmt.Println("received too few args:", os.Args)
+		fmt.Println("always invoke this via kustomize plugins")
+		os.Exit(1)
+	}
+
+	// ignore the first file name argument
+	// load the second argument, the file path
+	content, err := ioutil.ReadFile(os.Args[1])
+
+	if err != nil {
+		fmt.Println("unable to read in manifest", os.Args[1])
+		os.Exit(1)
+	}
+
+	var manifest ksops
+	err = yaml.Unmarshal(content, &manifest)
+
+	if err != nil {
+		fmt.Printf("error unmarshalling manifest content: %q \n%s\n", err, content)
+		os.Exit(1)
+	}
+
+	if manifest.Files == nil {
+		fmt.Println("missing the required 'files' key in the ksops manifests")
+		os.Exit(1)
+	}
+
+	var output bytes.Buffer
+
+	for _, file := range manifest.Files {
+		b, err := ioutil.ReadFile(file)
+
+		if err != nil {
+			fmt.Printf("error reading %q: %q\n", file, err.Error())
+			os.Exit(1)
+		}
+
+		format := formats.FormatForPath(file)
+		data, err := decrypt.DataWithFormat(b, format)
+
+		if err != nil {
+			fmt.Printf("trouble decrypting file %s", err.Error())
+			os.Exit(1)
+		}
+
+		output.Write(data)
+		output.WriteString("\n---\n")
+	}
+
+	fmt.Print(output.String())
+}
